When possible, call $(MAKE) rather than $(MAKEREC)
MAKEREC is a make variable that allows one to call make with the
'-f Makefile.nt' option.

When the main build system builds something in a subdirectory where
Makefile and Makefile.nt have been merged, it can now do that by using
$(MAKE) rather than $(MAKEREC).

This commit implements this change. For the root directory, where the
makefiles have not yet been merged, the calls to $(MAKEREC) have been
kept.