Make testsuite able to run ocamltest-based tests
This commit enhances the test harness so that it becomes able to
run ocamltest-based tests besides to the legacy tests it already runs.

This is achieved as follows:

Modification of targets in testsuite/Makefile:

- make legacy runs the legacy tests

- make new runs the ocamltest-based tests

  This uses ocamltest.opt if present and falls back to ocamltest
  (bytecode version) otherwise.

- make all runs both the legacy tests and the ocamltest-based tests

To know which tests to run, the testsuite/tests directory tree is
explored recursively as follows. For each of its subdirectory d:

- If d contains a Makefile, then d is assumed to contain legacy tests

- If d contains an ocamltests file, then d is assumed to contain
  ocamltest-based tests whose names are listed in the ocamltests file,
  one per line.

- If none of these files is found, then the subdirectories of d are explored
  recursively.