Migrate testsuite/tests/basic-more
The tests in this directory were compiled to native code using the
-inline 20 option. This behaviour has been preserved, except for the
morematch.ml. Indeed, for that test, it appears that for an inlining
level >=5, ocamlopt.byte and ocamlopt.opt produce different native
code binaries (the binaries are the ssame when the inlining level is <=4).

This was not an issue when the tests were run in the legacy way, but
it becomes one with ocamltest which is, by default, more strict.