Migrate testsuite/tests/basic
Before this commit, pr6322.ml was generated from a .in file. This was
done that way because the test is relevant only when OCaml is _not_
compiled with safe-string mode enabled.

This commit changes that: ocamltest provides a test that can be used
to skip further tests when safe-string mode is disabled.

This commit also removes the trigraph.ml.silent-compilation file.
It was here to mark that the compilation of the corresponding test (trigraph)
should not produce any output (see testsuite/makefiles/Makefile.several).

Such tricks are not needed any longer because it is the default behaviour
of ocamltest to ensure that compiling tests produces no output.

Moreover, on recent versons of OpenBSD, the linker produces a warning
which caused the compilation not to be silent as required.

This is why 3ac92c0ceef63fd1402ca8fb0db5fff3b469f88c skipped that test
on OpenBSD.

However, on recent OpenBSDs, ocamltest passes the -ccopt -nopie options
which avoid the warning and make compiling tests silent again.