Look for .git directory in srcdir rather than in the current working directory
When configure is called from the build directory, looking for
the .gitdirectory in the current working directory does not work
any longer, meaning that the computation of the C dependencies
is disabled.

This commit fixes this by looking for the .git directory in ${srcdir},
which should find it even when configure is called from the build tree.