Move the definitions of magic numbers to build-aux/ocaml_version.m4
This so so that the magic numbers can be used in several places in the
build system without having to be duplicated -- all the files that
use them can be generated by configure.

This commit does not touch the runtime because in some circumstances
like bootstrapping it is crucial that the magicnumbers in the compiler
and those in the runtime can be different.