Release the dependency of Meta on the debug_event concrete type
This commit uses a polymorphic type rather than Instruct.debug_event, which
releases the dependency of Meta on the Instruct module, which pulled
several other internal modules describing types involved in the
representation of debug_events.

Using an abstract deb_g_even type to keep some degree of type-safety
has been considered but it does not work because in
toplevel/bytecode/topeval.ml requires that the two types are the same.