Use dedicated types to represent comp units and predefs in CMO files
This commit introduces the dedicated `compunit` (resp. `predef`) types to
represent names of compilation units (resp. predefined exceptions) in
CMO files. This makes the CMO format independent of the type checker's
 `Ident.t` type which, although it is domain-specific, can represent
identifiers which can actually not occur in CMO files, namely the
Local and Scoped categories of identifiers.

This commit also adds a `Global` module to `Symtable` to represent the
fact that, for the time being, the symbol table contains both
compilation units and predefined exceptions.