Propagate the newly introduced types to the Lambda module
This commit continues to propagate the newly introduced types, this time
to the Lambda.primitive type.

In a similar way to what has been done in the Cmo_format module,
the Pgetglobal constructor is replaced by the Pgetpredef and Pgetcompunit
constructors, while Psetglobal becomes Psetcompunit, each of these
contructors now taking parameters of
types Cmo_format.compunit and Cmo_format.predef as appropriate.