Extend the OCAML_{BYTECODE,NATIVE}_PROGRAM macros to support C stubs
The macros assume that the C files that need to be compiled and linked to
build program foo are listed in the foo_STUBS variable, without their .c
extension.

The macros also assume that each  bar.c file has been compiled twice,
once as bar.b.o to be linked in bytecode mode and once as
bar.n.o tobe linked in natvie mode.