Make the build of ocamltest fit with the generic build framework
Currently, to build the program foo, the generic build framework
reads lists of files to use from the variables foo_LIBRARIES, foo_STUBS
and foo_MODULES.

However, ocamltest's build system rather starts by listing all the
source files (no matter their kind) in one variable and derives
all the information it needs from that.

This commit introduces the variables expected by the generic build
infrastructure but derives them from the list of source files.

Most of the per-file type varialbes used by ocamltest's build system
have been preserved because it could be good to update the generic
build infrastructure so that it works the same way ocamltest's build
system works currently. We would keep foo_LIBRARIES but start from
foo_SOURCES and derive foo_MODULES and foo_STUBS.

By doing so, one would for example become able to pass precise lists
of files to ocamldep when computing dependencies, rather than wildcards.

The install targets could take advantage of this approach, too, because
it would become possible to describe in an abstract way what it
means to install a program or library. For example, installing the
bytecode version of library foo can be specified as: install foo.cma
and $(foo_CMI_FILES), $(foo_MLI_FILES).

One side effect of this commit is that the C files used in ocamltest are
now built twice: once to be linked in the bytecode version and once
for linking in the native version.

This is why we had to add rules to Makefile.common but it may be the case
that once ocamltest/Makefile is merged into the root Makefile the rules
in Makefile.common are no longer needed because those in
the root Makefile almost cover these files, except that at the moment
the rules in the root Makefile apply only to files in the runtime directory
so they should be made more generic.