In Makefile.common, also keep track of configured modules
It is unclear yet whether we should keep track of source files, their
associated build arterfacts or both.

There is also the problem of how the macros in Makefile.common can be
expanded over several lines, if we have to do so or want to.

The variables introduced to track lexers, parsers and configured
moduels should be used in clean and distclean targets to remove the
generated files.

We will need to make sure our system still works in the ocntext of
out-of-source builds.